+++ src/main/java/com/codetaylor/mc/artisanworktables/modules/worktables/ModuleWorktables.java
@@ -102,6 +102,7 @@ public class ModuleWorktables
         TileEntityWorktableJeweler.class,
         TileEntityWorktableMage.class,
         TileEntityWorktableMason.class,
+        TileEntityWorktableScribe.class,
         TileEntityWorktableTailor.class
     );
   }